requirements.txt
langgraph>=0.2.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
